pyproject.toml
[tool.poetry]
name = "fastapiauthservice"
version = "0.1.0"
description = ""
authors = ["denisahmetov <denis.akhmetov.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
